01-Tuesday-Exercises/Customers/src/main/java/dbfacade/CustomerFacade.java
package dbfacade;

import entity.Customer;
import java.util.ArrayList;
import java.util.List;
import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import javax.persistence.Persistence;
import javax.persistence.Query;
import javax.persistence.TypedQuery;

/**
 *
 * @author Mathias
 */
public class CustomerFacade {

    private static EntityManagerFactory emf;
    private static CustomerFacade instance;

    private CustomerFacade() {
    }

    public static void main(String[] args) {
        EntityManagerFactory emf = Persistence.createEntityManagerFactory("pu");
        CustomerFacade facade = CustomerFacade.getCustomerFacade(emf);
        Customer c1 = facade.addCustomer("Mathias", "Nielsen");
        Customer c2 = facade.addCustomer("Nicklas", "Nielsen");
        Customer c3 = facade.addCustomer("Nikolaj", "Larsen");
        
        System.out.println("Customer1: " + facade.findByID(c1.getId()).getFirstName());
        System.out.println("Customer2: " + facade.findByID(c2.getId()).getFirstName());
        System.out.println("Customer3: " + facade.findByID(c3.getId()).getFirstName());
        
        System.out.println("Number of customers: " + facade.getNumberOfCustomers());

    }

    public static CustomerFacade getCustomerFacade(EntityManagerFactory _emf) {
        if (instance == null) {
            emf = _emf;
            instance = new CustomerFacade();
        }
        return instance;
    }

    public Customer findByID(int id) {
        EntityManager em = emf.createEntityManager();
        try {
            Customer customer = em.find(Customer.class, id);
            return customer;
        } finally {
            em.close();
        }

    }

    public List<Customer> findByLastName(String name){
        EntityManager em = emf.createEntityManager();
        try{
            TypedQuery<Customer> query = em.createQuery("SELECT c FROM Customer c WHERE c.lastName = :lastname",Customer.class);
            query.setParameter("lastname", name);
            List<Customer> customer = query.getResultList();
            return customer;
        }finally{
            em.close();
        }
    }

    public int getNumberOfCustomers() {
        EntityManager em = emf.createEntityManager();
        try {
            Query query = em.createQuery("SELECT COUNT(c) FROM Customer c", Customer.class);
            long customers = (long) query.getSingleResult();
            int i = (int) customers;
            return i;
        } finally {
            em.close();
        }
    }

    public List<Customer> allCustomers() {
        EntityManager em = emf.createEntityManager();
        try {
            TypedQuery<Customer> query = em.createQuery("SELECT c FROM Customer c", Customer.class);
            return query.getResultList();
        } finally {
            em.close();
        }
    }

    public Customer addCustomer(String fName, String lName) {
        Customer customer = new Customer(fName, lName);
        EntityManager em = emf.createEntityManager();
        try {
            em.getTransaction().begin();
            em.persist(customer);
            em.getTransaction().commit();
            return customer;
        } finally {
            em.close();
        }

    }
}
